requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
